pyproject.toml
[project]
name = "fertpy"
version = "0.3.0"
description = "Domain-driven deterministic agronomic calculation engine"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Luiz Henrique de Lima Peguin" },
    { name = "Pedro Henrique Escaranaro Brasil" }
]
keywords = ["agronomy", "fertilizer", "soil", "liming", "agriculture"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "PyYAML>=6.0",
    "numpy>=1.24",
    "scipy>=1.10"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2"
]

[project.urls]
Homepage = "https://github.com/lhpeguin/Fertpy"
Repository = "https://github.com/lhpeguin/Fertpy"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fertpy = ["knowledge/**/*.yaml"]